Support 64 bit UEFI live boot for created and rebuilt images.
This is more of a hack. I've found this comment in debian forums and
then I've just followed the dots.

<http://forums.debian.net/viewtopic.php?f=20&t=127380#p609579>

Which then leads to James Bottomley blog:
<http://blog.hansenpartnership.com/linux-foundation-secure-boot-system-released/>.

Thank you for this, James Bottomley and "Head_on_a_Stick"
(<http://forums.debian.net> and <https://forums.bunsenlabs.org/>). It
would not be possible without you guys.